requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.4.0
billiard==4.2.1
black==25.1.0
build==1.3.0
celery==5.4.0
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-config-url==0.1.1
dj-database-url==2.3.0
Django==5.2.3
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-stubs==5.2.1
django-stubs-ext==5.2.1
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.10
kombu==5.4.2
magic-filter==1.0.12
msgpack==1.1.0
multidict==6.7.0
mypy==1.16.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.4.1
psycopg2==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
pyproject_hooks==1.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==6.2.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
yarl==1.22.0
